Django/optify/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
aioredis==1.3.1
tornado==6.0.4
asyncpg==0.20.1

cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.3.0
google-api-core[grpc]==1.16.0
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-auth==1.13.1
google-cloud-language==1.3.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.2
idna==2.9
joblib==0.14.1
django-filter==2.2.0
kiwisolver==1.2.0
matplotlib==3.2.1
nltk==3.5
numpy==1.18.2
oauthlib==3.1.0
pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests-oauthlib==1.3.0
requests==2.23.0
rest-framework-generic-relations==2.0.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.25.8
